+++ src/rootfilespec/bootstrap/TBasket.py
@@ -1,0 +1,96 @@
+"""TBasket seems to be assumed in ROOT files and is not in the TStreamerInfo"""
+
+from typing import Annotated
+
+import numpy as np
+from numpy.typing import NDArray
+
+from rootfilespec.bootstrap.TKey import TKey
+from rootfilespec.bootstrap.TObject import StreamHeader
+from rootfilespec.dispatch import DICTIONARY
+from rootfilespec.structutil import (
+    Fmt,
+    ReadBuffer,
+    ROOTSerializable,
+    serializable,
+)
+
+
+@serializable
+class TBasket_header(ROOTSerializable):
+    fVersion: Annotated[int, Fmt(">h")]
+    "Version of the TBasket class"
+    fBufferSize: Annotated[int, Fmt(">i")]
+    "fBuffer length in bytes"
+    fNevBufSize: Annotated[int, Fmt(">i")]
+    "Length in Int_t of fEntryOffset OR fixed length of each entry if fEntryOffset is null!"
+    fNevBuf: Annotated[int, Fmt(">i")]
+    "Number of entries in basket"
+    fLast: Annotated[int, Fmt(">i")]
+    "Pointer to last used byte in basket"
+    flag: Annotated[int, Fmt(">B")]
+    """Some flags that control what is read/written:
+        - flag % 10 !=2 means fEntryOffset is available
+    """
+    # fIOBits: Annotated[int, Fmt(">B")]
+    # "IO feature flags.  Serialized in custom portion of streamer to avoid forward compat issues unless needed."
+
+
+@serializable
+class TBasket(TKey):
+    """TBasket object"""
+
+    bheader: TBasket_header
+    fEntryOffset: NDArray[np.int32]  #  BasicArray(np.dtype(">i"), "fNevBuf")
+    "Offset of entries in fBuffer(TKey)"
+    fBuffer: bytes
+    "Buffer if the basket owns it"
+
+    @classmethod
+    def read(cls, buffer: ReadBuffer):
+        args, buffer = cls.read_members(buffer)
+        return cls(*args), buffer
+
+    @classmethod
+    def read_members(cls, buffer: ReadBuffer):
+        start_position = buffer.relpos
+        sheader, buffer = StreamHeader.read(buffer)
+        (header, fSeekKey, fSeekPdir, fClassName, fName, fTitle), buffer = (
+            TKey.read_members(buffer)
+        )
+        bheader, buffer = TBasket_header.read(buffer)
+
+        fEntryOffset: NDArray[np.int32] = np.full(
+            bheader.fNevBuf, -1, dtype=np.dtype(">i")
+        )
+        fBuffer = b""
+        # TODO: full flag handling
+        # https://github.com/root-project/root/blob/0e6282a641b65bdf5ad832882e547ca990e8f1a5/tree/tree/src/TBasket.cxx#L993-L1027
+        if bheader.flag not in (11, 12):
+            msg = f"TBasket header flag {bheader.flag} not supported"
+            raise ValueError(msg)
+        if bheader.flag % 10 != 2:
+            # TODO: refactor this to use BasicArray
+            (n,), buffer = buffer.unpack(">i")
+            assert n == bheader.fNevBuf
+            data, buffer = buffer.consume(n * fEntryOffset.dtype.itemsize)
+            fEntryOffset = np.frombuffer(data, dtype=fEntryOffset.dtype, count=n)
+        if header.is_embedded():
+            _, buffer = TKey.read_members(buffer)
+            _, buffer = TBasket_header.read(buffer)
+            end_position = start_position + sheader.fByteCount + 4
+            fBuffer, buffer = buffer.consume(end_position - buffer.relpos)
+        return (
+            header,
+            fSeekKey,
+            fSeekPdir,
+            fClassName,
+            fName,
+            fTitle,
+            bheader,
+            fEntryOffset,
+            fBuffer,
+        ), buffer
+
+
+DICTIONARY["TBasket"] = TBasket
