+++ src/rootfilespec/rntuple/__init__.py
@@ -1,7 +1,7 @@
 """RNTuple class definitions
 
 This module contains the class definitions for the RNTuple format.
-The format schema is definied in the ROOT documentation:
+The format schema is defined in the ROOT documentation:
 https://root.cern/doc/v634/md_tree_2ntuple_2v7_2doc_2BinaryFormatSpecification.html
 
 Generally, all primitives in this module are encoded little-endian.
